+++ C++/DirichletTaskSolver/matrix.h
@@ -13,13 +13,13 @@ class Matrix {
  public:
   Matrix() { clear(); }
 
-  Matrix(int num_rows, int num_cols, const T& default_value = T())
+  Matrix(size_t num_rows, size_t num_cols, const T& default_value = T())
     : num_rows_(num_rows)
     , num_cols_(num_cols)
   {
     std::vector<T> temp(num_cols_, default_value);
 
-    for (int i = 0; i < num_rows; ++i) {
+    for (size_t i = 0; i < num_rows; ++i) {
       data_.push_back(temp);
     }
   }
@@ -33,15 +33,15 @@ class Matrix {
     data_.clear();
   }
 
-  int num_rows() const { return num_rows_; }
-  int num_cols() const { return num_cols_; }
+  size_t num_rows() const { return num_rows_; }
+  size_t num_cols() const { return num_cols_; }
 
-  T& operator()(int row_index, int col_index) { return data_[row_index][col_index]; }
-  const T& operator()(int row_index, int col_index) const { return data_[row_index][col_index]; }
+  T& operator()(size_t row_index, size_t col_index) { return data_[row_index][col_index]; }
+  const T& operator()(size_t row_index, size_t col_index) const { return data_[row_index][col_index]; }
 
   void debug_print() const {
-    for (int i = 0; i < num_rows(); ++i) {
-      for (int j = 0; j < num_cols(); ++j) {
+    for (size_t i = 0; i < num_rows(); ++i) {
+      for (size_t j = 0; j < num_cols(); ++j) {
         std::cout << data_[i][j] << " ";
       }
       std::cout << std::endl;
@@ -58,8 +58,8 @@ class Matrix {
     }
     auto retval = std::shared_ptr<Matrix<T> >(new Matrix(num_rows(), num_cols(), 0.0));
 
-    for (int i = 0; i < num_rows(); ++i) {
-      for (int j = 0; j < num_cols(); ++j) {
+    for (size_t i = 0; i < num_rows(); ++i) {
+      for (size_t j = 0; j < num_cols(); ++j) {
         (*retval)(i, j) = data_[i][j] + rhs(i, j);
       }
     }
@@ -76,8 +76,8 @@ class Matrix {
     }
     auto retval = std::shared_ptr<Matrix<T> >(new Matrix(num_rows(), num_cols(), 0.0));
 
-    for (int i = 0; i < num_rows(); ++i) {
-      for (int j = 0; j < num_cols(); ++j) {
+    for (size_t i = 0; i < num_rows(); ++i) {
+      for (size_t j = 0; j < num_cols(); ++j) {
         (*retval)(i, j) = data_[i][j] - rhs(i, j);
       }
     }
@@ -94,8 +94,8 @@ class Matrix {
     }
     auto retval = std::shared_ptr<Matrix<T> >(new Matrix(num_rows(), num_cols(), 0.0));
 
-    for (int i = 0; i < num_rows(); ++i) {
-      for (int j = 0; j < num_cols(); ++j) {
+    for (size_t i = 0; i < num_rows(); ++i) {
+      for (size_t j = 0; j < num_cols(); ++j) {
         (*retval)(i, j) = data_[i][j] * rhs(i, j);
       }
     }
@@ -105,8 +105,8 @@ class Matrix {
   std::shared_ptr<Matrix<T> > operator*(const T& val) const {
     auto retval = std::shared_ptr<Matrix<T> >(new Matrix(num_rows(), num_cols(), 0.0));
 
-    for (int i = 0; i < num_rows(); ++i) {
-      for (int j = 0; j < num_cols(); ++j) {
+    for (size_t i = 0; i < num_rows(); ++i) {
+      for (size_t j = 0; j < num_cols(); ++j) {
         (*retval)(i, j) = data_[i][j] * val;
       }
     }
@@ -116,8 +116,8 @@ class Matrix {
   static std::shared_ptr<Matrix<T> > FivePointsLaplass(const Matrix<T>& src, const Grid& grid) {
     auto retval = std::shared_ptr<Matrix<T> >(new Matrix<T>(src.num_rows(), src.num_cols(), 0.0));
 
-    for (int i = 1; i < retval->num_rows() - 1; ++i) {
-      for (int j = 1; j < retval->num_cols() - 1; ++j) {
+    for (size_t i = 1; i < retval->num_rows() - 1; ++i) {
+      for (size_t j = 1; j < retval->num_cols() - 1; ++j) {
         double part_1 = (src(i, j) - src(i - 1, j)) / grid.step_height() -
                         (src(i + 1, j) - src(i, j)) / grid.step_height();
         double part_2 = (src(i, j) - src(i, j - 1)) / grid.step_width() -
@@ -136,8 +136,8 @@ class Matrix {
       throw std::runtime_error("Matrix::ProductByPointAndSum: inconsistent sizes of matrices");
     }
 
-    for (int i = 0; i < src_1.num_cols(); ++i) {
-      for (int j = 0; j < src_2.num_rows(); ++j) {
+    for (size_t i = 0; i < src_1.num_cols(); ++i) {
+      for (size_t j = 0; j < src_2.num_rows(); ++j) {
         retval += grid.step_height() * grid.step_width() * src_1(i, j) * src_2(i, j);
       }
     }
@@ -154,8 +154,8 @@ class Matrix {
     data_ = rhs.data_;
   }
 
-  int num_rows_;
-  int num_cols_;
+  size_t num_rows_;
+  size_t num_cols_;
   std::vector<std::vector<T> > data_;
 };
